Extract Presenter
