+++ Spring-Boot/Spring-Boot-Student/src/main/java/tw/waterball/judgegirl/springboot/student/controllers/AdminController.java
@@ -16,16 +16,15 @@ package tw.waterball.judgegirl.springboot.student.controllers;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import tw.waterball.judgegirl.commons.token.TokenService;
-import tw.waterball.judgegirl.entities.Student;
+import tw.waterball.judgegirl.springboot.student.presenters.GetStudentsPresenter;
+import tw.waterball.judgegirl.springboot.student.presenters.SignInPresenter;
+import tw.waterball.judgegirl.springboot.student.presenters.SignUpPresenter;
 import tw.waterball.judgegirl.springboot.student.view.StudentView;
 import tw.waterball.judgegirl.studentservice.domain.usecases.GetStudentsWithFilterUseCase;
 import tw.waterball.judgegirl.studentservice.domain.usecases.SignInUseCase;
 import tw.waterball.judgegirl.studentservice.domain.usecases.SignUpUseCase;
 
 import java.util.List;
-import java.util.stream.Collectors;
-
-import static tw.waterball.judgegirl.springboot.student.view.StudentView.toViewModel;
 
 /**
  * @author chaoyulee chaoyu2330@gmail.com
@@ -42,14 +41,15 @@ public class AdminController {
 
     @PostMapping
     public StudentView signUp(@RequestBody SignUpUseCase.Request request) {
-        SignUpAdminPresenter presenter = new SignUpAdminPresenter();
-        signUpUseCase.execute(request, true, presenter);
+        SignUpPresenter presenter = new SignUpPresenter();
+        request.admin = true;
+        signUpUseCase.execute(request, presenter);
         return presenter.present();
     }
 
     @PostMapping("/login")
     public LoginResponse login(@RequestBody SignInUseCase.Request request) {
-        SignInAdminPresenter presenter = new SignInAdminPresenter();
+        SignInPresenter presenter = new SignInPresenter();
         signInUseCase.execute(request, presenter);
         presenter.setToken(tokenService.createToken(new TokenService.Identity(presenter.getStudentId())));
         return presenter.present();
@@ -59,60 +59,8 @@ public class AdminController {
     public List<StudentView> getAdminsWithFilter(
             @RequestParam(defaultValue = "0", required = false) int skip,
             @RequestParam(defaultValue = "25", required = false) int size) {
-        GetAdminsPresenter presenter = new GetAdminsPresenter();
+        GetStudentsPresenter presenter = new GetStudentsPresenter();
         getStudentsWithFilterUseCase.execute(new GetStudentsWithFilterUseCase.Request(skip, size, true), presenter);
         return presenter.present();
-    }
-}
-
-class SignUpAdminPresenter implements SignUpUseCase.Presenter {
-    private Student student;
-
-    @Override
-    public void setStudent(Student student) {
-        this.student = student;
-    }
-
-    StudentView present() {
-        return toViewModel(student);
-    }
-}
-
-class SignInAdminPresenter implements SignInUseCase.Presenter {
-    private Student student;
-    private TokenService.Token token;
-
-    @Override
-    public void setToken(TokenService.Token token) {
-        this.token = token;
-    }
-
-    @Override
-    public void setStudent(Student student) {
-        this.student = student;
-    }
-
-    Integer getStudentId() {
-        return student.getId();
-    }
-
-    LoginResponse present() {
-        return new LoginResponse(student.getId(), student.getEmail(), token.toString(),
-                token.getExpiration().getTime(), student.isAdmin());
-    }
-}
-
-class GetAdminsPresenter implements GetStudentsWithFilterUseCase.Presenter {
-    private List<Student> students;
-
-    @Override
-    public void setStudents(List<Student> students) {
-        this.students = students;
-    }
-
-    List<StudentView> present() {
-        return students.stream()
-                .map(StudentView::toViewModel)
-                .collect(Collectors.toList());
     }
 }
